fixup! Initial version of scripts for Enable OEM exclusive apps support feature
