+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/002_2_VehicleTypeData_without_not_mandatory_params_Ack.lua
@@ -21,7 +21,7 @@ local vehicleTypeData = {
 
 --[[ Local Functions ]]
 local function startRpcService(pAckParams)
-  local excludedParams = { "make","model", "modelYear", "trim", "systemHardwareVersion" }
+  local excludedParams = { "make","model", "modelYear", "trim" }
   common.startRpcService(pAckParams)
   :ValidIf(function(_, data)
     local errorMessages = ""
