+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/003_EndService_right_after_StartServiceAck.lua
@@ -5,7 +5,7 @@
 --  the vehicle type data received from SDL and requests EndService after StartServiceAck
 --
 -- Steps:
--- 1. HMI provides all vehicle type data in BC.GetSystemInfo(ccpu_version, systemHardwareVersion)
+-- 1. HMI provides all vehicle type data in BC.GetSystemInfo(ccpu_version)
 --  and VI.GetVehicleType(make, model, modelYear, trim) responses
 -- 2. App requests StartService(RPC) via 5th protocol
 -- SDL does:
