+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_2_GetVT_no_response_initial_cap.lua
@@ -6,17 +6,17 @@
 --  to VI.GetVehicleType request
 --
 -- Steps:
--- 1. HMI provides BC.GetSystemInfo(ccpu_version, systemHardwareVersion)
+-- 1. HMI provides BC.GetSystemInfo(ccpu_version)
 -- 2. HMI does not respond to VI.GetVehicleType requests
 -- 3. App requests StartService(RPC) via 5th protocol
 -- SDL does:
---  - Provide systemSoftwareVersion and systemHardwareVersion values received from HMI in BC.GetSystemInfo response
+--  - Provide systemSoftwareVersion value received from HMI in BC.GetSystemInfo response
 --     via StartServiceAck to the app
 --  - Provide the values for make, model, modelYear, trim parameters from the initial SDL capabilities file defined in
 --     .ini file in HMICapabilities parameter via StartServiceAck to the app
 -- 4. App sends RAI request via 5th protocol
 -- SDL does:
---  - Provide systemSoftwareVersion and systemHardwareVersion values received from HMI in BC.GetSystemInfo response
+--  - Provide systemSoftwareVersion value received from HMI in BC.GetSystemInfo response
 --     via RAI response to the app
 --  - Provide the values for make, model, modelYear, trim parameters from the initial SDL capabilities file defined in
 --     .ini file in HMICapabilities parameter via RAI response to the app
@@ -32,8 +32,7 @@ local vehicleTypeInfoParams = {
   model = defaultVTD.model,
   modelYear = defaultVTD.modelYear,
   trim = defaultVTD.trim,
-  ccpu_version = customVTD.ccpu_version,
-  systemHardwareVersion = customVTD.systemHardwareVersion
+  ccpu_version = customVTD.ccpu_version
 }
 
 --[[ Local Functions ]]
