+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/010_VehicleTypeData_RAI_4th_protocol.lua
@@ -5,7 +5,7 @@
 --  via 4th protocol
 --
 -- Steps:
--- 1. HMI provides all vehicle type data in BC.GetSystemInfo(ccpu_version, systemHardwareVersion)
+-- 1. HMI provides all vehicle type data in BC.GetSystemInfo(ccpu_version)
 --  and VI.GetVehicleType(make, model, modelYear, trim) responses
 -- 2. App requests RAI via 4th protocol
 -- SDL does:
