+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/011_1_VehicleTypeData_from_DB_and_updated_from_GetVT.lua
@@ -1,23 +1,23 @@
 ---------------------------------------------------------------------------------------------------
 -- Proposal: https://github.com/smartdevicelink/sdl_evolution/blob/master/proposals/0293-vehicle-type-filter.md
 ---------------------------------------------------------------------------------------------------
--- Description: SDL is able to provide systemSoftwareVersion and systemHardwareVersion versions from the DB
---  and updated values from VI.GetVehicleType response in the second SDL ignition cycle
+-- Description: SDL is able to provide systemSoftwareVersion version from the DB and updated values from
+--  VI.GetVehicleType response in the second SDL ignition cycle
 --
 -- Steps:
 -- 1. HMI responds with erroneous code to BC.GetSystemInfo request in the second ignition cycle,
---  systemSoftwareVersion and systemHardwareVersion have been saved to the DB in the previous ignition cycle
+--  systemSoftwareVersion has been saved to the DB in the previous ignition cycle
 -- SDL does:
 --  - Remove the cache file with hmi capabilities
 --  - Request obtaining of all HMI capabilities and VI.GetVehicleType RPC
 -- 2. HMI responds with updated values to VI.GetVehicleType request
 -- 3. App requests StartService(RPC) via 5th protocol
 -- SDL does:
---  - Provide systemHardwareVersion and systemSoftwareVersion values from the DB in StartServiceAck to the app
+--  - Provide systemSoftwareVersion value from the DB in StartServiceAck to the app
 --  - Provide the values for make, model, modelYear, trim parameters received from HMI in StartServiceAck to the app
 -- 4. App requests RAI
 -- SDL does:
---  - Provide systemHardwareVersion and systemSoftwareVersion values from the DB in RAI response to the app
+--  - Provide systemSoftwareVersion value from the DB in RAI response to the app
 --  - Provide the values for make, model, modelYear, trim parameters received from HMI in RAI response to the app
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
@@ -29,30 +29,12 @@ local vehicleTypeInfoParams = {
   model = common.vehicleTypeInfoParams.custom.model,
   modelYear = common.vehicleTypeInfoParams.custom.modelYear,
   trim = common.vehicleTypeInfoParams.custom.trim,
-  ccpu_version = common.vehicleTypeInfoParams.default.ccpu_version,
-  systemHardwareVersion = common.vehicleTypeInfoParams.default.systemHardwareVersion
+  ccpu_version = common.vehicleTypeInfoParams.default.ccpu_version
 }
 
 local defaultHmiCap = common.setHMIcap(common.vehicleTypeInfoParams.default)
 
 --[[ Local Functions ]]
-local function getRpcServiceAckParams(pVehicleTypeInfoParams)
-  local ackParams = {
-    make = common.setStringBsonValue(pVehicleTypeInfoParams.make),
-    model = common.setStringBsonValue(pVehicleTypeInfoParams.model),
-    modelYear = common.setStringBsonValue(pVehicleTypeInfoParams.modelYear),
-    trim = common.setStringBsonValue(pVehicleTypeInfoParams.trim),
-    systemSoftwareVersion = common.setStringBsonValue(pVehicleTypeInfoParams.ccpu_version),
-    systemHardwareVersion = common.setStringBsonValue(pVehicleTypeInfoParams.systemHardwareVersion)
-  }
-  for key, KeyValue in pairs(ackParams) do
-    if not KeyValue.value then
-      ackParams[key] = nil
-    end
-  end
-  return ackParams
-end
-
 local function startErrorResponseGetSystemInfo()
   local hmiCap = common.setHMIcap(common.vehicleTypeInfoParams.custom)
   hmiCap.BasicCommunication.GetSystemInfo = nil
@@ -72,7 +54,7 @@ common.Step("Ignition off", common.ignitionOff)
 common.Title("Test")
 common.Step("Start SDL, HMI sends GetSystemInfo(GENERIC_ERROR) response", startErrorResponseGetSystemInfo )
 common.Step("Start RPC Service, Vehicle type data in StartServiceAck",
-  common.startRpcService, { getRpcServiceAckParams(vehicleTypeInfoParams) })
+  common.startRpcService, { common.getRpcServiceAckParamsFromStruct(vehicleTypeInfoParams) })
 common.Step("Vehicle type data in RAI", common.registerAppEx, { vehicleTypeInfoParams })
 
 common.Title("Postconditions")
