+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_1_GetSI_ccpu_updated_and_updated_from_GetVT.lua
