+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/014_All_VehicleTypeData_pretected_Ack.lua
@@ -4,7 +4,7 @@
 -- Description: Check that SDL is able to provide vehicle type data in the encrypted StartServiceAck
 --
 -- Steps:
--- 1. HMI provides all vehicle type data in BC.GetSystemInfo(ccpu_version, systemHardwareVersion)
+-- 1. HMI provides all vehicle type data in BC.GetSystemInfo(ccpu_version)
 --  and VI.GetVehicleType(make, model, modelYear, trim) responses
 -- 2. Unprotected RPC service is started
 -- 3. App requests protected StartService(RPC) via 5th protocol
