+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/015_2_GetSI_and_GetVT_delayed_multiple_Apps.lua
