+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_2_GetVT_delayed_RAI_on_4th_protocol.lua
