+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_3_GetVT_GetSI_delayed_RAI_on_4th_protocol.lua
