+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_4_GetSI_new_ign_cycle_delayed_RAI_on_4th_protocol.lua
@@ -16,7 +16,7 @@
 -- 4. HMI responds with delay to BC.GetSystemInfo request
 -- SDL does:
 --  - Send RAI response after receiving BC.GetSystemInfo response
---  - Provide systemSoftwareVersion and systemHardwareVersion values received from HMI in BC.GetSystemInfo response
+--  - Provide systemSoftwareVersion value received from HMI in BC.GetSystemInfo response
 --     via RAI response to the app
 --  - Provide the values for make, model, modelYear, trim parameters from the cache in RAI response to the app
 ---------------------------------------------------------------------------------------------------
