+++ test_sets/oem_exclusive.txt
@@ -9,21 +9,21 @@
 ./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_1_GetVT_error_initial_cap.lua
 ./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_2_GetVT_no_response_initial_cap.lua
 ./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_3_GetVT_invalid_resp_initial_cap.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/008_1_GetSI_delayed.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/008_2_GetVT_delayed.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/008_3_GetSI_and_GetVT_delayed.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/008_4_GetSI_delayed_new_Ign_Cycle.lua
 ./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/009_Cached_VehicleTypeData_Ack_RAI.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/010_GetSI_systemHardwareVersion_snapshot.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/011_GetSI_systemHardwareVersion_valid_boundary.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_1_GetSI_systemHardwareVersion_invalid_boundary_empty_values.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_2_GetSI_systemSoftwareVersion_invalid_boundary_value_from_DB.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/013_VehicleTypeData_RAI_4th_protocol.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/014_1_VehicleTypeData_from_DB_and_updated_from_GetVT.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/014_2_VehicleTypeData_from_DB_and_from_initial_cap.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/015_1_GetSI_ccpu_updated_and_updated_from_GetVT.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/015_2_GetSI_ccpu_version_updated_and_from_initial_cap.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_Audio_video_services.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/017_All_VehicleTypeData_pretected_Ack.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/018_1_VehicleTypeData_multiple_apps.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/018_2_GetSI_and_GetVT_delayed_multiple_Apps.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/019_1_GetSI_delayed_RAI_on_4th_protocol.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/019_2_GetVT_delayed_RAI_on_4th_protocol.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/019_3_GetVT_GetSI_delayed_RAI_on_4th_protocol.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/019_4_GetSI_new_ign_cycle_delayed_RAI_on_4th_protocol.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/010_VehicleTypeData_RAI_4th_protocol.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/011_1_VehicleTypeData_from_DB_and_updated_from_GetVT.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/011_2_VehicleTypeData_from_DB_and_from_initial_cap.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_1_GetSI_ccpu_updated_and_updated_from_GetVT.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_2_GetSI_ccpu_version_updated_and_from_initial_cap.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/013_Audio_video_services.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/014_All_VehicleTypeData_pretected_Ack.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/015_1_VehicleTypeData_multiple_apps.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/015_2_GetSI_and_GetVT_delayed_multiple_Apps.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_1_GetSI_delayed_RAI_on_4th_protocol.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_2_GetVT_delayed_RAI_on_4th_protocol.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_3_GetVT_GetSI_delayed_RAI_on_4th_protocol.lua
+./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_4_GetSI_new_ign_cycle_delayed_RAI_on_4th_protocol.lua
